+++ app/views/index/my.php
@@ -6,6 +6,10 @@
 
 <?= $this->render_partial('index/_confirm_delete') ?>
 
+<?php if (empty($meetings)): ?>
+	<?= MessageBox::info(_('Es sind keine Meetings vorhanden. Meetings können nur innerhalb einer Veranstaltung über das "+"-Icon aktiviert und anschließend über den Reiter "Meetings" verwaltet werden.')) ?>
+<?php endif; ?>
+
 <?php if ($type === 'name'): ?>
     <?=$this->render_partial('index/_meetings', array('title' => _('Meine Meetings'), 'canModifyMeetings' => true, 'destination' => 'index/my/name', 'showCourse' => true)) ?>
 <?php else: ?>