+++ CHANGELOG.md
@@ -2,3 +2,14 @@
 =====
 
 * use guzzle/guzzle to perform API HTTP requests
+
+
+1.1.0
+=====
+
+* supports feature-tours and context-sensitive help-content
+* fixing utf8-encoding bug in DfnVcDriver.php
+* changes navigation-position of "my" from header to profile
+* sets plugin-title to "Meetings (Betatest)"
+* adds MessageBox with helping information in /views/index/my.php
+